Shader fixes
Do signed preservation in shader instead of using snorm textures as it's
not a requirement for OpenGL to allow snorm textures to be drawable.

This is a workaround for a bug in the nvidia driver which reports that
the snorm texture attached correctly to fbo. Ideally if it reported that
it doesn't work it should then switch between the two methods, but since
nvidia doesn't report that it doesn't work then we have to assume that it
doesn't work.